+++ src/main/java/model/AttackStatus.java
@@ -1,5 +1,10 @@
 package model;
 
+/**
+ * Enum for outcomes of fight against other ranks in a game.
+ * Used to determine the result of an attack: WON, LOST, DRAW, FLAG_CAPTURE, or ERROR.
+ * 
+ */
 public enum AttackStatus {
 	
 	WON, LOST, DRAW, FLAG_CAPTURE, ERROR;